requirements.txt
pillow==11.2.1
pygame==2.6.1
pynetworktables==2021.0.0
